requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
sqlalchemy>=1.4.46
redis>=4.5.4
python-dotenv>=0.19.2
aiosmtplib>=1.1.6
apscheduler>=3.6.3
passlib>=1.7.4
python-multipart>=0.0.5
openai>=1.0.0
jinja2>=3.0.0
cryptography>=41.0.0
pymysql>=1.1.0
PyJWT>=2.8.0
bcrypt>=4.0.0
websockets>=12.0
loguru>=0.7.0
blackboxprotobuf>=1.0.0
